extract send function into other file
